requirements.txt
discord.py==2.5.2
httpx==0.28.1
apscheduler==3.11.0
motor==3.7.1
openai==1.98.0
python-dotenv==1.1.1
pytz==2025.2
finnhub-python==2.4.24
pydantic==2.11.7
websockets==15.0.1
yfinance==0.2.65